Add basic test from all endpoints
The majority of this commit is adding the required functions and
fixtures to conftests for the test cases to achieve the intended
outcome. The added test cases simply test the basic expected happy path
of the requirements, next steps for them would be to parametrize those
happy paths for further verification followed by the additional of cases
to cover fault paths.